+++ mods/src/multiplayer/internal.h
@@ -3,3 +3,4 @@
 __attribute__((visibility("internal"))) void _init_multiplayer_raknet();
 __attribute__((visibility("internal"))) void _init_multiplayer_loading();
 __attribute__((visibility("internal"))) void _init_multiplayer_syncing();
+__attribute__((visibility("internal"))) void _init_multiplayer_inventory();